freecad: Ignore global gitconfig
If a global /etc/gitconfig exists, git inside the sandbox will attempt to read it.
However, this is forbidden and thus git inside the snap fails with:

warning: unable to access '/etc/gitconfig': Permission denied
fatal: unknown error occurred while reading the configuration files

As a result, the addon manager gets stuck at the "Cloning module..." stage.
Prevent this by telling git to ignore the global config, which is not necessary at all
to download the addons.

Signed-off-by: Maximilian Federle <max.federle@gmail.com>